+++ src/Attributes/Strings/ReplaceLast.php
@@ -1,0 +1,33 @@
+<?php declare(strict_types=1);
+
+/**
+ * Copyright (C) Brian Faust
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Cline\Struct\Attributes\Strings;
+
+use Attribute;
+use Illuminate\Support\Str;
+
+/**
+ * Replaces the last matching value in a string property.
+ *
+ * @author Brian Faust <brian@cline.sh>
+ * @psalm-immutable
+ */
+#[Attribute(Attribute::TARGET_PROPERTY)]
+final readonly class ReplaceLast extends AbstractStringTransformer
+{
+    public function __construct(
+        public string $search,
+        public string $replace,
+    ) {}
+
+    public function transform(string $value): string
+    {
+        return Str::replaceLast($this->search, $this->replace, $value);
+    }
+}