+++ src/Attributes/Strings/Squish.php
@@ -1,0 +1,28 @@
+<?php declare(strict_types=1);
+
+/**
+ * Copyright (C) Brian Faust
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Cline\Struct\Attributes\Strings;
+
+use Attribute;
+use Illuminate\Support\Str;
+
+/**
+ * Collapses internal whitespace and trims the outer edges of a string property.
+ *
+ * @psalm-immutable
+ * @author Brian Faust <brian@cline.sh>
+ */
+#[Attribute(Attribute::TARGET_PROPERTY)]
+final readonly class Squish extends AbstractStringTransformer
+{
+    public function transform(string $value): string
+    {
+        return Str::squish($value);
+    }
+}