+++ src/Attributes/Strings/Ulid.php
@@ -1,0 +1,28 @@
+<?php declare(strict_types=1);
+
+/**
+ * Copyright (C) Brian Faust
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Cline\Struct\Attributes\Strings;
+
+use Attribute;
+use Illuminate\Support\Str;
+
+/**
+ * Generates a ULID when the input key is missing.
+ *
+ * @author Brian Faust <brian@cline.sh>
+ * @psalm-immutable
+ */
+#[Attribute(Attribute::TARGET_PROPERTY)]
+final readonly class Ulid extends AbstractGeneratedValueAttribute
+{
+    public function generate(): string
+    {
+        return $this->normalize((string) Str::ulid());
+    }
+}